Phase 2: multi-moment decoder (VEL, ZDR, PHI, RHO, SW)
- Radial.data upgraded from [UInt8] to [UInt16] to cover 16-bit NEXRAD
  moments (ZDR, PHI, VEL use 16-bit; REF, RHO use 8-bit — word size read
  from data block header byte 19)
- Level2Decoder now parses all supported 'D' type blocks per radial instead
  of stopping at REF; returns one RadarSweep per (elevation, momentType) pair
- AppState caches allSweeps after first download; selectCurrentSweep() picks
  from the cache by product + tilt so product/tilt changes don't re-download
- ProductPicker onChange wired up (was previously a no-op)
- TiltPicker onChange uses selectCurrentSweep() instead of re-downloading
- RadarOverlay dispatches to product-specific color tables:
  velocity (blue/green/red), ZDR, correlation coefficient, differential
  phase (hue wheel), and spectrum width
- New integration test verifies REF + ZDR + PHI + RHO all decode from KEWX

+++ wxaccess/NEXRAD/RadarSweep.swift
@@ -22,7 +22,9 @@ struct Radial: Sendable {
     let numGates: Int
     let scale: Float           // raw → physical: physical = (raw - offset) / scale
     let offset: Float
-    let data: [UInt8]          // raw gate values; 0 = below threshold, 1 = range folded
+    // Gate values: 0 = below threshold, 1 = range folded, 2+ = valid.
+    // UInt16 covers both 8-bit (REF, RHO) and 16-bit (VEL, ZDR, PHI) NEXRAD moments.
+    let data: [UInt16]
 
     func physicalValue(gateIndex: Int) -> Float? {
         guard gateIndex < data.count else { return nil }
